rideshare-trip-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pg8000==1.30.4
redis==5.0.1
prometheus-client==0.19.0
geopy==2.4.0
python-dotenv==1.0.0
python-dateutil==2.9.0
flask-sock
flask-swagger-ui
flasgger==0.9.7.1 